Adds some validations and completes flows

+++ src/products/use-cases/delete-product/deleteProductUseCase.ts
@@ -1,5 +1,7 @@
 import { IProductRepo } from '../../ports/out/productRepo';
 import { DeleteProductDTO } from './deleteProductDTO';
+import { Result } from '../../../shared/core/Result';
+import { GeneralErrors } from '../../../shared/core/GeneralErrors';
 
 export class DeleteProductUseCase {
   private productRepo: IProductRepo;
@@ -9,6 +11,22 @@ export class DeleteProductUseCase {
   }
 
   async execute(dto: DeleteProductDTO) {
-    return await this.productRepo.delete(dto);
+    try {
+      if (!dto.id) {
+        return Result.fail(new GeneralErrors().BadRequest());
+      }
+
+      const product = await this.productRepo.readById(dto);
+
+      if (!product) {
+        return Result.fail(new GeneralErrors().NotFound());
+      }
+
+      await this.productRepo.delete(dto);
+
+      return Result.ok(dto);
+    } catch (error) {
+      return Result.fail(new GeneralErrors().InternalServerError());
+    }
   }
 }
